% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uslopes.R
\name{USlopes}
\alias{USlopes}
\title{Estimates slopes at extremes}
\usage{
USlopes(lmObject, vars, .vcov = NULL, x.min = NULL, x.max = NULL)
}
\arguments{
\item{lmObject}{The model to be tested}

\item{vars}{A vector with the name of the linear and squared terms}

\item{.vcov}{The covariance matrix to use}

\item{x.min}{Lower bound of interval}

\item{x.max}{Upper bound of interval}
}
\description{
This function estimates the slope of the relationship at the extreme points of the independent variable
}
