% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utest-package.R
\docType{package}
\name{utest-package}
\alias{utest}
\alias{utest-package}
\title{utest: Implementation of the Lind/Mehlum Utest}
\description{
An implementation of Lind and Mehlum's (2010) Utest to test for the presence of a U shaped or inverted U shaped relationship between variables in (generalized) linear models. It also implements a test of upward/downward sloping relationships at the lower and upper boundary of the data range.
}
\author{
\strong{Maintainer}: Jo Thori Lind \email{j.t.lind@econ.uio.no} (\href{https://orcid.org/0000-0002-2980-3814}{ORCID})

}
\keyword{internal}
