% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utest.R
\name{UTest}
\alias{UTest}
\title{Perform Utest}
\usage{
UTest(lmObject, vars, .vcov = NULL)
}
\arguments{
\item{lmObject}{The model to be tested}

\item{vars}{A vector with the name of the linear and squared terms}

\item{.vcov}{The covariance matrix to use}

\item{x.min}{Lower bound of interval}

\item{x.max}{Upper bound of interval}
}
\description{
This function computes the Lind/Mehlum test of a U shaped relationship.
}
